Items into the items, take action and code correctnes
Now an item can be inside an other item.
A few items more had been created.
A review of the code correctnes has been
done. Now you can take an item. The
function checks if the player already have
the item, else check if the room have the item.

An item can have a pointer to an other
item which is a must to collect him.