Exits locekd and restriction
Now the exists can have be closed.
If they are closed they can have a
pointer to an entity and depends of
the type  of this entity the exit will have
a restriction to unlock.

The hall room with the exit connections
has been added to the game. The door
is closed to go from "The door -> The hall"
To open the door player must have more
than 75% of his HP.